flask_test1/requirements.txt
Flask
Flask-CORS
Flask-SQLAlchemy
PyMySQL
bcrypt
PyJWT
python-dotenv